Moving transformer to utilities folder
