+++ Starter.WebApi.IntegrationTests/GlobalUsings.cs
@@ -1,1 +1,3 @@
-﻿global using Starter.WebApi.Models.Entities;
+﻿global using Starter.WebApi.IntegrationTests.Facts.Factories;
+global using Starter.WebApi.Models.Authentication;
+global using Starter.WebApi.Models.Entities;
