+++ Starter.WebApi.UnitTests/GlobalUsings.cs
@@ -6,7 +6,7 @@ global using Microsoft.Extensions.Logging;
 global using Moq;
 global using Starter.WebApi.Controllers;
 global using Starter.WebApi.Models.Authentication;
-global using Starter.WebApi.Models.Database;
+global using Starter.WebApi.Models.Entities;
 global using Starter.WebApi.Models.DataTransferObjects;
 global using Starter.WebApi.Models.DataTransferObjects.Profiles;
 global using Starter.WebApi.Services;
