+++ Starter.WebApi/Controllers/Abstracts/StarterControllerBase.cs
@@ -1,6 +1,4 @@
-﻿using System.Text.RegularExpressions;
-
-namespace Starter.WebApi.Controllers.Abstracts;
+﻿namespace Starter.WebApi.Controllers.Abstracts;
 
 /// <summary>
 /// Abstraction for application controllers
@@ -67,25 +65,4 @@ public class StarterControllerBase(IMapper mapper) : ControllerBase
 
         return Ok(mappedValue);
     }
-}
-
-/// <summary>
-/// Use kebab case for route
-/// </summary>
-public partial class ToKebabParameterTransformer : IOutboundParameterTransformer
-{
-    /// <summary>
-    /// Convert pascal case to kebab case
-    /// </summary>
-    /// <param name="value">Initial route value</param>
-    /// <returns>Route with kebab case as naming convention</returns>
-    public string TransformOutbound(object? value)
-    {
-        return MatchLowercaseThenUppercase()
-            .Replace(value?.ToString() ?? "", "$1-$2")
-            .ToLower();
-    }
-
-    [GeneratedRegex("([a-z])([A-Z])")]
-    private static partial Regex MatchLowercaseThenUppercase();
 }
